// fem-vision-studio/design-mockup/src/lib/types/timber/CltPreset.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Voorinstelling: een gangbare symmetrische opbouw, vrij aanpasbaar.
 * De lagen wisselen af lengte/dwars/lengte/…, beginnend met een lengtelaag.
 */
export type CltPreset = { 
/**
 * Naam zoals in de keuzelijst, bijv. "5-laags 160".
 */
name: string, 
/**
 * Laagdikten van boven naar beneden in mm.
 */
thicknesses_mm: Array<number>, 
/**
 * Totale dikte in mm.
 */
height_mm: number, };